Hero, Research, and Sponsor Update
Centralize data in /constants/index.js for al the arrays used in the infiniteMenu and BounceCards components. Update infiniteMenu component to redirect the user to the project page when clicked the central item, use the titles and descriptions as the projectCards component, and resize the text font and width. Update the CircularGallery component to dynamically manage the resize of the title text when its bigger than the scale width. Finally, updated the BounceCards component to redirect the user to the sponsor's website when the card is clicked, and change the display image from object-cover to object-constrain so it doesn't overflow.